Move alias functionality in own file
